+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 1.6.0
+
+* Added `@babel/plugin-transform-named-capturing-groups-regex`.
+* Added `@babel/plugin-transform-dotall-regex`.
+* Removed CoreJS v2 support.
+* Updated packages to latest version.
+* Bumped babel requirement to 7.6.0 or higher.
+* Added dependency scanning to the build script.
+
 ## 1.5.0
 
 * Deprecated CoreJS v2, next major version will only use CoreJS ^3.